API-3245 | update nodejs client
