+++ .npmignore
@@ -1,0 +1,3 @@
+test
+samples.js
+config.example.json
