+++ lib/constants/index.js
@@ -1,6 +1,8 @@
 module.exports = {
   BASE_URL: 'https://api.pdffiller.com',
   APPLICATIONS_ENDPOINT: '/v2/applications',
+  APPLICATIONS_BY_ID_ENDPOINT: '/v2/applications/{id}',
+  APPLICATIONS_USER_BY_ID_ENDPOINT: '/v2/applications/{id}/users',
   AUTH_ENDPOINT: '/v2/oauth/token',
   TOKENS_ENDPOINT: '/v2/tokens',
   USERS_ENDPOINT: '/v2/users/me',
