+++ lib/extends/base.js
@@ -1,0 +1,9 @@
+const makeRequest = require('../api').makeRequest;
+
+class BaseExtend {
+  constructor() {
+    this.request = makeRequest;
+  }
+}
+
+module.exports = BaseExtend;
